+++ _test/test/help_test.dart
@@ -10,35 +10,35 @@ import 'package:test/test.dart';
 import 'common/utils.dart';
 
 void main() {
-  test('pub run build_runner help', () async {
+  test('dart run build_runner help', () async {
     await _testHelpCommand(['help'],
         checkContent: 'Unified interface for running Dart builds.');
     await _testHelpCommand(['--help'],
         checkContent: 'Unified interface for running Dart builds.');
   });
 
-  test('pub run build_runner build --help', () async {
+  test('dart run build_runner build --help', () async {
     await _testHelpCommand(['build', '--help'],
         checkContent: 'Performs a single build on the specified targets');
     await _testHelpCommand(['help', 'build'],
         checkContent: 'Performs a single build on the specified targets');
   });
 
-  test('pub run build_runner serve --help', () async {
+  test('dart run build_runner serve --help', () async {
     await _testHelpCommand(['serve', '--help'],
         checkContent: 'Runs a development server');
     await _testHelpCommand(['help', 'serve'],
         checkContent: 'Runs a development server');
   });
 
-  test('pub run build_runner test --help', () async {
+  test('dart run build_runner test --help', () async {
     await _testHelpCommand(['test', '--help'],
         checkContent: 'and then runs tests using');
     await _testHelpCommand(['help', 'test'],
         checkContent: 'and then runs tests using');
   });
 
-  test('pub run build_runner watch --help', () async {
+  test('dart run build_runner watch --help', () async {
     await _testHelpCommand(['watch', '--help'],
         checkContent: 'watching the file system for updates');
     await _testHelpCommand(['help', 'watch'],